Разведены сериализеры и вью по разным файлам
