+++ backend/api/serializers/instrument_serializers.py
@@ -1,0 +1,8 @@
+from job.models import Instrument
+
+
+class InstrumentSerializer(serializers.ModelSerializer):
+    """Сериализатор для модели Instrument."""
+    class Meta:
+        model = Instrument
+        fields = '__all__'
