+++ backend/api/serializers/language_serializers.py
@@ -1,0 +1,7 @@
+from job.models import Language
+
+
+class LanguageSerializer(serializers.ModelSerializer):
+    class Meta:
+        model = Language
+        fields = '__all__'
