+++ backend/api/serializers/skill_serializers.py
@@ -1,0 +1,8 @@
+from job.models import Skill
+
+
+class SkillSerializer(serializers.ModelSerializer):
+    """Сериализатор для модели Skill."""
+    class Meta:
+        model = Skill
+        fields = '__all__'
