+++ backend/api/serializers/specialization_serializers.py
@@ -1,0 +1,7 @@
+from job.models import Specialization
+
+
+class SpecializationSerializer(serializers.ModelSerializer):
+    class Meta:
+        model = Specialization
+        fields = '__all__'
