+++ backend/api/serializers/sphere_serializers.py
@@ -1,0 +1,8 @@
+from job.models import Sphere
+
+
+class SphereSerializer(serializers.ModelSerializer):
+
+    class Meta:
+        model = Sphere
+        fields = '__all__'
