+++ backend/api/views/caseimage_views.py
@@ -1,0 +1,13 @@
+from api.serializers.caseimage_serializers import CaseImageSerializer
+from job.models import CaseImage
+
+
+class CaseImageViewSet(viewsets.ModelViewSet):
+    http_method_names = ('get', 'post')
+    queryset = CaseImage.objects.all()
+    serializer_class = CaseImageSerializer
+
+    @action(detail=True,
+            methods=['get', ])
+    def portfolio(self, request, pk):        
+        cover = CaseImage
