+++ backend/api/views/skill_views.py
@@ -1,0 +1,10 @@
+from rest_framework.permissions import AllowAny
+
+from api.serializers.skill_serializers import SkillSerializer
+from job.models import Skill
+
+class SkillViewSet(viewsets.ReadOnlyModelViewSet):
+    queryset = Skill.objects.all()
+    serializer_class = SkillSerializer
+    pagination_class = None
+    permission_classes = [AllowAny, ]
