+++ backend/api/views/specialization_views.py
@@ -1,0 +1,7 @@
+from api.serializers.specialization_serializers import SpecializationSerializer
+from job.models import Specialization
+
+
+class SpecializationViewSet(viewsets.ReadOnlyModelViewSet):
+    serializer_class = SpecializationSerializer
+    queryset = Specialization.objects.all()
